// student-growth-system-design-v3/student-growth-system-design-v3/frontend/src/api/generated/models/WrongStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Student Growth Archive API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 3.15.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const WrongStatus = {
    New: 'NEW',
    Reviewing: 'REVIEWING',
    Mastered: 'MASTERED',
    Archived: 'ARCHIVED'
} as const;
export type WrongStatus = typeof WrongStatus[keyof typeof WrongStatus];


export function instanceOfWrongStatus(value: any): boolean {
    for (const key in WrongStatus) {
        if (Object.prototype.hasOwnProperty.call(WrongStatus, key)) {
            if (WrongStatus[key as keyof typeof WrongStatus] === value) {
                return true;
            }
        }
    }
    return false;
}

export function WrongStatusFromJSON(json: any): WrongStatus {
    return WrongStatusFromJSONTyped(json, false);
}

export function WrongStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): WrongStatus {
    return json as WrongStatus;
}

export function WrongStatusToJSON(value?: WrongStatus | null): any {
    return value as any;
}

export function WrongStatusToJSONTyped(value: any, ignoreDiscriminator: boolean): WrongStatus {
    return value as WrongStatus;
}